moved test_exception_on_debug over to TestMySQLHub, where it belongs
